Mag field cache move to double

See merge request atlas/athena!33754

+++ MagneticField/MagFieldElements/MagFieldElements/BFieldCacheZR.h
@@ -66,8 +66,8 @@ public:
 private:
   double m_zmin, m_zmax; // bin range in z
   double m_rmin, m_rmax; // bin range in r
-  float m_invz, m_invr;  // 1/(bin size) in z, r
-  float m_field[2][4];   // (Bz,Br) at 4 corners of the bin
+  double m_invz, m_invr;  // 1/(bin size) in z, r
+  double m_field[2][4];   // (Bz,Br) at 4 corners of the bin
 };
 #include "BFieldCacheZR.icc"
 #endif
